// app/src/api/table/TableClient.ts
export interface TableEvent {
  id: string;
  author: string;
  event_type: string;
  timestamp: number;
  content: string;
  ability: string | null;
  d20_roll: number | null;
  modifier: number | null;
  total: number | null;
}